backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
sendgrid==6.11.0
httpx==0.26.0
beautifulsoup4==4.12.3
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
stripe==7.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
